5.Line_bot/requirements.txt
line-bot-sdk
flask
imutils
cmake
dlib
opencv-python
keras
tensorflow
opencv-python

